Update Stats records to match return type from DB

+++ src/main/java/de/htwberlin/casino/blackjack/application/domain/service/emitStats/OverviewStats.java
@@ -10,6 +10,6 @@ package de.htwberlin.casino.blackjack.application.domain.service.emitStats;
  */
 public record OverviewStats(Long totalGames,
                             Long totalPlayers,
-                            int totalBet,
-                            int houseProfit) implements Stats {
+                            double totalBet,
+                            double houseProfit) implements Stats {
 }
